pyproject.toml
[project]
name = "my-kafka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.49.1",
    "kafka-python>=2.2.15",
    "pytest>=8.4.2",
]
